move api to invtype. delete redundant files.
