+++ src/controllers/bmdashboard/bmInventoryTypeController.js
@@ -1,5 +1,5 @@
-const bmInventoryTypeController = function (InvType) {
-  const fetchMaterialTypes = async (req, res) => {
+function bmInventoryTypeController(InvType) {
+  async function fetchMaterialTypes(req, res) {
     try {
       InvType
         .find()
@@ -9,11 +9,22 @@ const bmInventoryTypeController = function (InvType) {
     } catch (err) {
       res.json(err);
     }
-  };
+  }
 
+  async function addEquipmentType(req, res) {
+    const { name, desc } = req.body;
+    console.log(name, desc);
+    try {
+      res.status(201).send({ message: 'Hello world!' });
+    } catch (error) {
+      res.status(500).send(error);
+    }
+  }
   return {
     fetchMaterialTypes,
+    addEquipmentType,
   };
-};
+}
+
 
 module.exports = bmInventoryTypeController;
