--- src/routes/bmdashboard/bmEquipmentRouter.js
@@ -1,12 +1,0 @@
-const express = require('express');
-
-const routes = function () {
-  const equipmentRouter = express.Router();
-  const controller = require('../../controllers/bmdashboard/bmEquipmentController')();
-  equipmentRouter.route('/equipment/add')
-    .post(controller.addEquipmentType);
-
-  return equipmentRouter;
-};
-
-module.exports = routes;
