+++ src/startup/routes.js
@@ -65,7 +65,6 @@ const bmLoginRouter = require('../routes/bmdashboard/bmLoginRouter')();
 const bmMaterialsRouter = require('../routes/bmdashboard/bmMaterialsRouter')(inventoryItemMaterial, buildingMaterial);
 const bmProjectRouter = require('../routes/bmdashboard/bmProjectRouter')(buildingProject);
 const bmInventoryTypeRouter = require('../routes/bmdashboard/bmInventoryTypeRouter')(buildingInventoryType);
-const bmEquipmentRouter = require('../routes/bmdashboard/bmEquipmentRouter')();
 
 module.exports = function (app) {
   app.use('/api', forgotPwdRouter);
@@ -102,5 +101,4 @@ module.exports = function (app) {
   app.use('/api/bm', bmMaterialsRouter);
   app.use('/api/bm', bmProjectRouter);
   app.use('/api/bm', bmInventoryTypeRouter);
-  app.use('/api/bm', bmEquipmentRouter);
 };
